Refactored jb::itch5::mold_udp_channel out of the tool into the library for testability
