+++ Makefile.am
@@ -46,6 +46,7 @@ jb_itch5_unit_tests = \
   jb/itch5/ut_message_header \
   jb/itch5/ut_mold_udp_pacer \
   jb/itch5/ut_mold_udp_pacer_config \
+  jb/itch5/ut_mold_udp_channel \
   jb/itch5/ut_mwcb_breach_message \
   jb/itch5/ut_mwcb_decline_level_message \
   jb/itch5/ut_net_order_imbalance_indicator_message \
@@ -285,6 +286,7 @@ jb_libitch5_a_SOURCES = \
   jb/itch5/ipo_quoting_period_update_message.cpp \
   jb/itch5/market_participant_position_message.cpp \
   jb/itch5/message_header.cpp \
+  jb/itch5/mold_udp_channel.cpp \
   jb/itch5/mold_udp_pacer_config.cpp \
   jb/itch5/mwcb_breach_message.cpp \
   jb/itch5/mwcb_decline_level_message.cpp \
@@ -319,6 +321,7 @@ jb_libitch5_a_HEADERS = \
   jb/itch5/ipo_quoting_period_update_message.hpp \
   jb/itch5/market_participant_position_message.hpp \
   jb/itch5/message_header.hpp \
+  jb/itch5/mold_udp_channel.hpp \
   jb/itch5/mold_udp_pacer.hpp \
   jb/itch5/mold_udp_pacer_config.hpp \
   jb/itch5/mold_udp_protocol_constants.hpp \
@@ -564,6 +567,10 @@ jb_itch5_ut_market_participant_position_message_LDADD = $(jb_ut_itch5_libs)
 jb_itch5_ut_message_header_SOURCES = jb/itch5/ut_message_header.cpp
 jb_itch5_ut_message_header_CPPFLAGS = $(UT_CPPFLAGS) -DBOOST_TEST_MODULE=jb_itch5_ut_message_header
 jb_itch5_ut_message_header_LDADD = $(jb_ut_itch5_libs)
+
+jb_itch5_ut_mold_udp_channel_SOURCES = jb/itch5/ut_mold_udp_channel.cpp
+jb_itch5_ut_mold_udp_channel_CPPFLAGS = $(UT_CPPFLAGS) -DBOOST_TEST_MODULE=jb_itch5_ut_mold_udp_channel
+jb_itch5_ut_mold_udp_channel_LDADD = $(jb_ut_itch5_libs)
 
 jb_itch5_ut_mold_udp_pacer_SOURCES = jb/itch5/ut_mold_udp_pacer.cpp
 jb_itch5_ut_mold_udp_pacer_CPPFLAGS = $(UT_CPPFLAGS) -DBOOST_TEST_MODULE=jb_itch5_ut_mold_udp_pacer
